requirements.txt
asgiref==3.6.0
Django==4.1.4
sqlparse==0.4.3
python-dotenv==0.21.0
psycopg2==2.9.5
Faker==15.3.4
djangorestframework==3.14.0
PyYAML==6.0
uritemplate==4.1.1
drf-yasg==1.21.4
django-filter==22.1
djangorestframework-simplejwt==5.2.2